+++ lib/src/types.dart
@@ -7,4 +7,8 @@ class LatLngData {
 
   /// Estimated horizontal accuracy, radial, in meters.
   final double accuracy;
+
+  bool highAccurency() {
+    return !(accuracy == null || accuracy <= 0.0 || accuracy > 30.0);
+  }
 }